refactor(json-crdt-extensions): 💡 strucutre data transfer code under /transfer folder
