+++ src/json-crdt-extensions/peritext/block/__tests__/Fragment-export-html.spec.ts
@@ -1,5 +1,5 @@
 import {type Kit, runAlphabetKitTestSuite} from '../../__tests__/setup';
-import {toHtml, toJsonMl} from '../../lazy/export-html';
+import {toHtml, toJsonMl} from '../../transfer/export-html';
 import {CommonSliceType} from '../../slice';
 
 const runTests = (setup: () => Kit) => {
