--- src/json-crdt-extensions/peritext/lazy/README.md
@@ -1,2 +1,0 @@
-Useful features which can be lazy-loaded (dynamically imported) separately from
-the main codebase.
