+++ src/json-crdt-extensions/peritext/transfer/README.md
@@ -1,0 +1,5 @@
+Useful data transfer utilities for Peritext. This module can be used to transfer
+export and import fragments of Peritext documents to and from other formats.
+
+Import and export modules can be lazy-loaded (dynamically imported) separately
+from the main codebase.
