+++ src/json-crdt-extensions/peritext/transfer/__tests__/PeritextDataTransfer.spec.ts
@@ -1,9 +1,9 @@
-import * as htmlExport from '../lazy/export-html';
-import * as htmlImport from '../lazy/import-html';
-import * as mdExport from '../lazy/export-markdown';
-import * as mdImport from '../lazy/import-markdown';
+import * as htmlExport from '../export-html';
+import * as htmlImport from '../import-html';
+import * as mdExport from '../export-markdown';
+import * as mdImport from '../import-markdown';
 import {PeritextDataTransfer} from '../PeritextDataTransfer';
-import {setupKit} from './setup';
+import {setupKit} from '../../__tests__/setup';
 
 const setup = () => {
   const kit = setupKit();
