+++ src/json-crdt-peritext-ui/events/index.ts
@@ -1,11 +1,11 @@
 import {PeritextEventDefaults, type PeritextEventDefaultsOpts} from './defaults/PeritextEventDefaults';
 import {PeritextEventTarget} from './PeritextEventTarget';
 import {DomClipboard} from './clipboard/DomClipboard';
-import {PeritextDataTransfer} from '../../json-crdt-extensions/peritext/PeritextDataTransfer';
-import * as htmlExport from '../../json-crdt-extensions/peritext/lazy/export-html';
-import * as htmlImport from '../../json-crdt-extensions/peritext/lazy/import-html';
-import * as mdExport from '../../json-crdt-extensions/peritext/lazy/export-markdown';
-import * as mdImport from '../../json-crdt-extensions/peritext/lazy/import-markdown';
+import {PeritextDataTransfer} from '../../json-crdt-extensions/peritext/transfer/PeritextDataTransfer';
+import * as htmlExport from '../../json-crdt-extensions/peritext/transfer/export-html';
+import * as htmlImport from '../../json-crdt-extensions/peritext/transfer/import-html';
+import * as mdExport from '../../json-crdt-extensions/peritext/transfer/export-markdown';
+import * as mdImport from '../../json-crdt-extensions/peritext/transfer/import-markdown';
 import type {Peritext} from '../../json-crdt-extensions';
 
 export const create = (txt: Peritext) => {
